feat: implement KelasDetailPage with presensi and materi management features
